feat: Add Java, Maven, and Quarkus support to dev container
This commit adds support for Java, Maven, and Quarkus to the dev container. It includes a new devcontainer-feature.json file that specifies the versions of Java, Maven, and Quarkus to install. Additionally, an install.sh script is added to install the specified versions of Java, Maven, and Quarkus using sdkman.